\name{get_couch}
\alias{get_couch}
\title{Get value from CouchDB storage based on key}
\usage{
get_couch(cache, key, db)
}
\arguments{
  \item{cache}{Logical}

  \item{key}{Key from url + args}

  \item{db}{Database name}
}
\description{
Get value from CouchDB storage based on key
}
\keyword{internal}

