\name{get_sqlite}
\alias{get_sqlite}
\title{Get value from SQlite storage based on key}
\usage{
get_sqlite(cache, key, db)
}
\arguments{
  \item{cache}{Logical}

  \item{key}{Key from url + args}

  \item{db}{Database name}
}
\description{
Get value from SQlite storage based on key
}
\keyword{internal}

