\name{goCache}
\alias{goCache}
\title{Search for data in a chosen backend}
\usage{
  goCache(cache, key, backend, path)
}
\arguments{
  \item{cache}{Logical}

  \item{key}{Key from url + args}

  \item{backend}{One of local, redis, couchdb, mongodb,
  sqlite.}
}
\description{
  Search for data in a chosen backend
}

