\name{make_key}
\alias{make_key}
\title{Make a key from a API call (base url for call, plus arguments).}
\usage{
make_key(url, args)
}
\arguments{
  \item{url}{base url for an API}

  \item{args}{A list with named arguments}
}
\description{
Make a key from a API call (base url for call, plus
arguments).
}
\keyword{internal}

